move testing functions to testing module
